Update README.md
# Segun

In my changes, I implement the video call with HTML5, CSS, Javascript and Agora SDK
In the project folder, I have index.html file which is the based url. From index.html, when user click on Peer Chat, it take them to one on one with with participant and when user clicks on create room button, it takes them to where they can create or join room by entering their name and room ID.

In one one call, user can decide to switch camera to off/on and same for mic

In room section, user can join room, switch camera to off/on, mic off/on, share screen as well, chat while in the room at right hand side, see participants from left sidebar, click on a circle video of participant to expand the video width, height and remove radius.

Users can leave call or room and participants will know the person that left.

I wrote this in HTML5, CSS and Javascript because of my busy schedule. If given more time, I can switch it to react.js